func for collecting birds

+++ common/qmath.h
@@ -6,7 +6,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-#define Q_UPSCALE(X,Y) ((int16_t)(((X)) << (Y)));
+#define Q_UPSCALE(X,Y) ((int16_t)(((uint16_t)(X)) << (Y)));
 #define Q_DNSCALE(X,Y) ((uint8_t)(((X)) >> (Y)));
 
 int16_t QMath_Mul(int16_t a, int16_t b, uint16_t q);
